[CD-192] Refactor: move load filelist func from dataset to utils
