+++ ai/sfen/train.py
@@ -17,11 +17,11 @@ from torch.nn.parallel import DistributedDataParallel
 from torch.utils.data import DistributedSampler, DataLoader
 
 
-from meldataset import MelDataset, mel_spectrogram, get_dataset_filelist
+from meldataset import MelDataset, mel_spectrogram
 from discriminator import MultiPeriodDiscriminator, MultiScaleDiscriminator
 from vae import VAE
 from loss import feature_loss, generator_loss, discriminator_loss, loss_function
-from utils import plot_spectrogram, scan_checkpoint, load_checkpoint, save_checkpoint
+from utils import plot_spectrogram, scan_checkpoint, load_checkpoint, save_checkpoint, get_dataset_filelist
 
 
 torch.backends.cudnn.benchmark = True
@@ -248,8 +248,8 @@ def main():
     parser.add_argument('--group_name', default=None)
     parser.add_argument('--input_wavs_dir', default='LJSpeech-1.1/wavs')
     parser.add_argument('--input_mels_dir', default='ft_dataset')
-    parser.add_argument('--input_training_file', default='LJSpeech-1.1/training.txt')
-    parser.add_argument('--input_validation_file', default='LJSpeech-1.1/validation.txt')
+    parser.add_argument('--input_training_file', default='filelists/train.txt')
+    parser.add_argument('--input_validation_file', default='filelists/val.txt')
     parser.add_argument('--checkpoint_path', default='cp_hifigan')
     parser.add_argument('--config', default='')
     parser.add_argument('--training_epochs', default=3100, type=int)
